[Feature] Add SGX unix socket detection support
In the current implementation, a new kubelet instance deletes all the existing
unix sockets under /var/lib/kubelet/device-plugins when it starts. Besides, the
SGX unix socket can be accidentally removed during ops.

To this end, we're proposing in this patch a detection mechanism for SGX device
plugin that will monitor the deletion of SGX unix socket, re-start the device
plugin grpc server and re-register itself with kubelet instance upon such an
event.

Signed-off-by: Kailun Qin <kailun.qkl@antfin.com>